+++ langchain/src/chains/tests/simple_sequential_chain.test.ts
@@ -3,7 +3,7 @@ import { BaseLLM } from "../../llms/base.js";
 import { LLMResult } from "../../schema/index.js";
 import { LLMChain } from "../llm_chain.js";
 import { PromptTemplate } from "../../prompts/index.js";
-import { SimpleSequentialChain } from "../simple_sequential_chain.js";
+import { SimpleSequentialChain } from "../sequential_chain.js";
 import { AnalyzeDocumentChain } from "../analyze_documents_chain.js";
 import { ConversationalRetrievalQAChain } from "../conversational_retrieval_chain.js";
 import { VectorStoreRetriever } from "../../vectorstores/base.js";
